becky - Added Becky.info (custom icon) - Added status line in GUI for last program action / key match - Added load and save of mappings from file - Added load and save of mappings from BEC - Added mouse-hover display of key mappings in both keyboards
FW
- Add services to get and set keymap entries
- Increased BEC message length to 2 bytes to eventually support larger messages
- Moved BEC message handling code to separate file (msg.c) for easier maintenance.
- Added fast path message handling for console input/output
  This allows the "bec term" command to not timeout on long commands.

+++ amiga/becmsg.h
@@ -12,10 +12,10 @@
 #ifndef _BECMSG_H
 #define _BECMSG_H
 
-uint send_cmd(uint8_t cmd, void *arg, uint8_t arglen,
+uint send_cmd(uint8_t cmd, void *arg, uint16_t arglen,
               void *reply, uint replymax, uint *replyalen);
 
-uint send_cmd_retry(uint8_t cmd, void *arg, uint8_t arglen,
+uint send_cmd_retry(uint8_t cmd, void *arg, uint16_t arglen,
                     void *reply, uint replymax, uint *replyalen);
 
 const char *bec_err(uint status);
